nros_params

Module typed

Typed parameter API

This module provides a fluent builder pattern for declaring typed parameters in a ROS 2 node, aligning with the rclrs API.

Re-exports

pub use crate::ParameterVariant;

Structs

MandatoryParameter

A parameter that must always have a value

OptionalParameter

A parameter that may or may not have a value

ParameterBuilder

Builder for declaring a typed parameter

ReadOnlyParameter

A parameter whose value cannot be changed after declaration

UndeclaredParameters

Provides access to undeclared parameters in a ParameterServer

Enums

ParameterError

Error type for typed parameter operations

Traits

RangeConvertible

Trait for types that can be converted to a parameter range from RangeInclusive